# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: K12 智学系统 API
  description: K12 智学系统 REST API 与 WebSocket 规范
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: 开发环境
  - url: https://api.example.com/api
    description: 生产环境
security:
  - bearerAuth: []
...
paths:
...
  /practice/judge:
    post:
      tags: [Practice]
      summary: 课堂练习判对错
      description: >-
        学生在 practice 卡片答题后调用。三路由判定：
        (1) 题库命中 + 客观题（choice/true_false/fill_blank）-> exact 答案比对；
        (2) 题库命中主观题（short_answer/proof）或未命中 -> AI JudgmentCapability 判定；
        (3) 答错 -> 写入 main_error_books（未入库的题先经 QuestionStructuringCapability
        结构化后插 questions 表，结构化失败则 question_id=NULL 仅存题面）。
        AI 判定失败返回 503（code=5001，不写错题本，前端可区分处理）。
      operationId: judgePractice
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JudgeRequest'
      responses:
        '200':
          description: 判定结果
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/CommonResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/JudgeResult'
        '503':
          description: AI 判定失败
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 5001
                  message:
                    type: string
                    example: 判定失败，请重试
  /practice/hint:
    post:
      tags: [Practice]
      summary: 课堂练习提示（AI 生成 + Card 级缓存）
      description: >-
        学生在 practice 卡片答题前点「提示」调用。先查 cards.hints 缓存（JSON：
        { "<题目文本>": "<提示文本>" }）命中直返（不调 AI）；未命中则调
        HintCapability 生成苏格拉底式提示（只启发不给答案）并写回 cards.hints，
        供后续复用（Card 级共享，不分学生）。AI 生成失败返回 503（code=5001，
        前端降级显示静态文案，不阻断答题）。
      operationId: getPracticeHint
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HintRequest'
      responses:
        '200':
          description: 提示内容
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/CommonResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/HintResult'
        '503':
          description: AI 提示生成失败
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 5001
                  message:
                    type: string
                    example: 提示生成失败，请重试
  /practice/discuss:
    post:
      tags: [Practice]
      summary: 课堂练习「让 AI 讲一讲」（苏格拉底讨论）
      description: >-
        学生在 practice 卡片答题前点「让 AI 讲一讲」调用。打开讨论即：① 记入主线错题本
        （find-or-create 幂等，source='discuss'，无论后续答对答错都保留）；② 创建带 card_id
        的 mainline 对话并返回 dialogueId。前端据此走 POST /ai/tutor/stream（mode=mainline）
        做苏格拉底式多轮讨论——TutoringCapability 据 card_id 解析 cardContent 限定范围（只能
        讨论该卡片内容，不能聊其它），3 次失败/放弃后兜底给完整解析。questionId 用
        content_hash 快查（不调 AI 结构化），未命中则 question_id=null + wrong_answer_text 存题面。
      operationId: startPracticeDiscuss
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiscussRequest'
      responses:
        '200':
          description: 错题本记录 + mainline 对话 dialogueId
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/CommonResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/DiscussResult'
        '500':
          description: 创建讨论会话失败
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 5000
                  message:
                    type: string
                    example: 创建讨论会话失败
  /practice/discuss-card:
    post:
      tags: [Practice]
      summary: 卡片级「思辨答疑」（苏格拉底讨论）
      description: >-
        学生在非 practice 知识卡片上点「思辨答疑」调用。与 `/practice/discuss` 区别：
        scope=整张卡片（非某道题），**不入错题本**（讨论的是知识而非一道题）。
        服务端按 (student, card_id, track='mainline') find-or-create 对话并返回 dialogueId；
        前端据此走 POST /ai/tutor/stream（mode=mainline）做苏格拉底式多轮讨论——
        TutoringCapability 据 card_id 解析 cardContent 限定范围（只能讨论该卡片内容）。
        重开同一卡片的讨论会自动回到同一对话线。
      operationId: startCardDiscuss
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiscussCardRequest'
      responses:
        '200':
          description: mainline 对话 dialogueId
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/CommonResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/DiscussCardResult'
        '500':
          description: 创建讨论会话失败
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 5000
                  message:
                    type: string
                    example: 创建讨论会话失败
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    CommonResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
      required: [code, message]
...
    JudgeRequest:
      type: object
      required: [cardId, lessonId, subjectId, questionText, studentAnswer]
      properties:
        cardId:
          type: integer
          description: 当前 practice 卡片 ID
        lessonId:
          type: integer
          description: 当前课程 ID
        subjectId:
          type: integer
          description: 学科 ID
        questionText:
          type: string
          description: 题面原文（取自 cards.content_metadata.questions[].text）
        studentAnswer:
          type: string
          description: 学生提交的答案文本
    JudgeResult:
      type: object
      properties:
        questionId:
          type: integer
          nullable: true
          description: 题库命中时的 questions.id；未命中且结构化失败时为 null（错题仅存题面）
        isCorrect:
          type: boolean
        method:
          type: string
          enum: [exact, ai]
          description: exact=客观题答案比对；ai=JudgmentCapability 判定
        analysis:
          type: string
          nullable: true
          description: 答错时的解析。method=ai 时为 AI 生成的分析；method=exact 答错时为标准答案文本
        errorType:
          type: string
          nullable: true
          enum: [logic, calculation, format, missing]
          description: AI 判定的错误类型，仅 method=ai 时返回
        errorBookId:
          type: integer
          nullable: true
          description: 答错时写入 main_error_books 的 id；答对或 AI 判定失败时不返回
    HintRequest:
      type: object
      required: [cardId, lessonId, subjectId, questionText]
      properties:
        cardId:
          type: integer
          description: 当前 practice 卡片 ID
        lessonId:
          type: integer
          description: 当前课程 ID
        subjectId:
          type: integer
          description: 学科 ID
        questionText:
          type: string
          description: 题面原文，同时作为 cards.hints JSON 的 key（即「题目标题」）
    HintResult:
      type: object
      properties:
        hint:
          type: string
          description: 苏格拉底式提示文本（只启发不给答案，公式用 $...$ LaTeX）
        cached:
          type: boolean
          description: true=命中 cards.hints 缓存直返（未调 AI）；false=本次 AI 新生成并已写回缓存
    DiscussRequest:
      type: object
      required: [cardId, lessonId, subjectId, questionText]
      properties:
        cardId:
          type: integer
          description: 当前 practice 卡片 ID（限定讨论范围 + 错题本 source_ref_id）
        lessonId:
          type: integer
          description: 当前课程 ID
        subjectId:
          type: integer
          description: 学科 ID
        questionText:
          type: string
          description: 题面原文；question_id 为空时作为错题本去重键（wrong_answer_text）
    DiscussResult:
      type: object
      properties:
        dialogueId:
          type: string
          description: mainline 对话 id，前端据此走 /ai/tutor/stream 做苏格拉底讨论
        errorBookId:
          type: integer
          description: 本次命中或新建的主线错题本记录 id
        questionId:
          type: integer
          nullable: true
          description: 题库中的题目 id（content_hash 未命中则为 null）
    DiscussCardRequest:
      type: object
      required: [cardId, lessonId, subjectId]
      properties:
        cardId:
          type: integer
          description: 当前知识卡片 ID（限定讨论范围）
        lessonId:
          type: integer
          description: 当前课程 ID
        subjectId:
          type: integer
          description: 学科 ID
    DiscussCardResult:
      type: object
      properties:
        dialogueId:
          type: string
          description: mainline 对话 id（find-or-create：可能复用已有讨论），前端据此走 /ai/tutor/stream 做苏格拉底讨论
...